Move all initializations utils&cruds to a specific utils
